requirements.txt

flask
flask-cors
openai
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
langchain
scikit-learn
nltk